Simplify directory validation
