+++ imessage-exporter/src/app/compatibility/converters/sticker.rs
@@ -2,7 +2,6 @@
  Defines routines for converting sticker image files.
 */
 
-
 use std::{
     fs::{create_dir_all, read_dir, remove_dir_all},
     path::{Path, PathBuf},
@@ -12,7 +11,7 @@ use imessage_database::tables::attachment::MediaType;
 
 use crate::app::compatibility::{
     converters::{
-        common::{copy_raw, run_command},
+        common::{copy_raw, ensure_paths, run_command},
         image::convert_heic,
     },
     models::{ImageConverter, ImageType, VideoConverter},
@@ -65,21 +64,7 @@ pub(crate) fn sticker_copy_convert(
 }
 
 fn convert_heics(from: &Path, to: &Path, video_converter: &VideoConverter) -> Option<()> {
-    // Get the path we want to copy from
-    let from_path = from.to_str()?;
-
-    // Get the path we want to write to
-    let to_path = to.to_str()?;
-
-    // Ensure the directory tree exists
-    if let Some(folder) = to.parent() {
-        if !folder.exists() {
-            if let Err(why) = create_dir_all(folder) {
-                eprintln!("Unable to create {folder:?}: {why}");
-                return None;
-            }
-        }
-    }
+    let (from_path, to_path) = ensure_paths(from, to)?;
 
     // Frames per second in the original sticker, generated by Apple
     let fps = 10;
